backend/pyproject.toml
[tool.poetry]
name = "personal-website"
version = "0.1.0"
description = "Personal website backend"
authors = ["zhiyuan"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.13.0,<4.0"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.35"
pydantic = "^2.11.0"
python-dotenv = "^1.1.0"
pymysql = "^1.1.1"
cryptography = "^45.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
